schedule-builder: move model to it own package to export
