+++ cmd/schedule-builder/cmd/root.go
@@ -24,6 +24,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
+	"k8s.io/release/cmd/schedule-builder/model"
 	"sigs.k8s.io/release-utils/log"
 	"sigs.k8s.io/yaml"
 )
@@ -121,8 +122,8 @@ func run(opts *options) error {
 	}
 
 	var (
-		patchSchedule   PatchSchedule
-		releaseSchedule ReleaseSchedule
+		patchSchedule   model.PatchSchedule
+		releaseSchedule model.ReleaseSchedule
 		scheduleOut     string
 	)
